requirements.txt
streamlit>=1.32.0
manim
elevenlabs
python-dotenv
flask
openai
mem0ai
